requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Deprecated==1.2.14
Django==5.1
django-otp==1.5.1
django-phone-verify==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
multidict==6.0.5
mysqlclient==2.2.4
nexmo==2.5.2
packaging==24.1
phonenumbers==8.13.43
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydotenv==0.0.7
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.2.0
sqlparse==0.5.1
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vonage==3.16.1
vonage-jwt==1.1.1
vonage-utils==1.1.2
wrapt==1.16.0
yarl==1.9.4
